revno: 2605 message:   Cannot startup patient-module.
committer: Tran Chau <tran.hispvietnam@gmail.com>
branch nick: dhis2
timestamp: Tue 2011-01-18 14:49:20 +0700

+++ dhis-2/dhis-services/dhis-service-patient/src/main/java/org/hisp/dhis/patient/startup/TableAlteror.java
@@ -67,34 +67,41 @@ public class TableAlteror
     public void execute()
         throws Exception
     {
-        updatePatientOrgunitAssociation();
+        try
+        {
+            updatePatientOrgunitAssociation( );
 
-        updateDOBType();
+            updateDOBType( );
 
-        executeSql( "UPDATE program SET version = 1 WHERE version is NULL" );
+            executeSql( "UPDATE program SET version = 1 WHERE version is NULL" );
 
-        updateDataSetMobileAttribute();
+            updateDataSetMobileAttribute();
 
-        updateDataSetVersionAttribute();
+            updateDataSetVersionAttribute();
 
-        executeSql( "UPDATE patientidentifiertype SET type='" + PatientIdentifierType.VALUE_TYPE_TEXT
-            + "' WHERE type IS NULL" );
+            executeSql( "UPDATE patientidentifiertype SET type='" + PatientIdentifierType.VALUE_TYPE_TEXT
+                + "' WHERE type IS NULL" );
 
-        executeSql( "UPDATE program SET minDaysAllowedInputData=0 WHERE minDaysAllowedInputData IS NULL" );
+            executeSql( "UPDATE program SET minDaysAllowedInputData=0 WHERE minDaysAllowedInputData IS NULL" );
 
-        executeSql( "UPDATE program SET maxDaysAllowedInputData=0 WHERE maxDaysAllowedInputData IS NULL" );
-        
-        executeSql( "UPDATE patient SET isdead=false WHERE isdead IS NULL" );
+            executeSql( "UPDATE program SET maxDaysAllowedInputData=0 WHERE maxDaysAllowedInputData IS NULL" );
+
+            executeSql( "UPDATE patient SET isdead=false WHERE isdead IS NULL" );
+
+        }
+        catch ( Exception ex )
+        {
+            log.error( ex );
+        }
 
     }
 
     private void updatePatientOrgunitAssociation()
     {
-
-        StatementHolder holder = statementManager.getHolder();
-
         try
         {
+            StatementHolder holder = statementManager.getHolder();
+
             Statement statement = holder.getStatement();
 
             ResultSet isUpdated = statement
@@ -116,70 +123,49 @@ public class TableAlteror
         }
         catch ( Exception ex )
         {
-            log.error( ex );
-        }
-        finally
-        {
-            holder.close();
+            log.debug( ex );
         }
     }
 
-    private void updateDOBType()
+    private void updateDOBType( )
+        throws Exception
     {
-        StatementHolder holder = statementManager.getHolder();
-
         try
         {
-            executeSql( "UPDATE patient SET dobType='A' WHERE birthdateestimated=true" );
+            StatementHolder holder = statementManager.getHolder();
+            Statement statement = holder.getStatement();
 
-            executeSql( "ALTER TABLE patient drop column birthdateestimated" );
+            ResultSet isUpdated = statement
+                .executeQuery( "SELECT * FROM INFORMATION_SCHEMA.COLUMNS WHERE TABLE_NAME = 'patient' AND COLUMN_NAME = 'birthdateestimated'" );
 
-            executeSql( "DELETE FROM validationcriteria where property='birthdateestimated'" );
+            if ( isUpdated.next() )
+            {
+                executeSql( "UPDATE patient SET dobType='A' WHERE birthdateestimated=true" );
+
+                executeSql( "ALTER TABLE patient drop column birthdateestimated" );
+
+                executeSql( "DELETE FROM validationcriteria where property='birthdateestimated'" );
+            }
+
+            executeSql( "UPDATE patient SET dobType='D' dobType is null" );
         }
         catch ( Exception ex )
         {
-            log.error( ex );
-        }
-        finally
-        {
-            holder.close();
+            log.debug( ex );
+
         }
     }
 
     private void updateDataSetMobileAttribute()
     {
-        StatementHolder holder = statementManager.getHolder();
+        executeSql( "UPDATE dataset SET mobile = false WHERE mobile is null" );
 
-        try
-        {
-            executeSql( "UPDATE dataset SET mobile = false WHERE mobile is null" );
-        }
-        catch ( Exception ex )
-        {
-            log.error( ex );
-        }
-        finally
-        {
-            holder.close();
-        }
     }
 
     private void updateDataSetVersionAttribute()
     {
-        StatementHolder holder = statementManager.getHolder();
+        executeSql( "UPDATE dataset SET version = 1 WHERE version is null" );
 
-        try
-        {
-            executeSql( "UPDATE dataset SET version = 1 WHERE version is null" );
-        }
-        catch ( Exception ex )
-        {
-            log.error( ex );
-        }
-        finally
-        {
-            holder.close();
-        }
     }
 
     private int executeSql( String sql )
